+++ src/renderer/components/Main.vue
@@ -4,7 +4,7 @@
     <v-container fill-height>
       <v-layout align-center row>
         <v-flex xs1>
-          <img src="../assets/logo.png"></img>
+          <img src="../assets/logo.png"/>
         </v-flex>
         <v-flex text-xs-center>
           <h3 id="main-app-name" class="display-3">Sixpence</h3>
@@ -21,6 +21,9 @@
         </v-btn>
         <v-btn id="main-expense-button" large color="red" @click="$router.push(`/expenses`)">
           <v-icon>mdi-currency-usd</v-icon> Expenses
+        </v-btn>
+        <v-btn id="main-report-button" large color="orange lighten-2" @click="$router.push(`/report/list`)">
+          <v-icon>mdi-file-chart</v-icon> Reports
         </v-btn>
       </v-flex>
     </v-layout>