+++ src/renderer/router/index.js
@@ -21,6 +21,16 @@ export default new Router({
       component: require('@/components/Expenses').default
     },
     {
+      path: '/report/list',
+      name: 'report-list',
+      component: require('@/components/Report/Main').default
+    },
+    {
+      path: '/report/yearly',
+      name: 'report-yearly',
+      component: require('@/components/Report/Yearly').default
+    },
+    {
       path: '*',
       redirect: '/'
     }