+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v1.5.0 (Unreleased)
+* New Feature: __Reports__
+    * Button on Home Page to navigate to Reports
+    * Reports added to View menu
+    * One report added "Yearly by Category"
+
 ## v1.4.0 (July 06, 2018)
 * Budget: Ability to show budgeted Income/Expenses for each month in a list.